# backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.CreateShiftRequest:
    properties:
      end_time:
        type: string
      note:
        type: string
      start_time:
        type: string
      zone_tag_id:
        type: integer
    required:
    - end_time
    - start_time
    type: object
...
  domain.PatchShiftRequest:
    properties:
      end_time:
        type: string
      note:
        type: string
      start_time:
        type: string
      zone_tag_id:
        type: integer
    type: object
...
  response.DeleteShiftResponse:
    properties:
      id:
        example: 1
        type: integer
      request_id:
        example: 7fbd6854-8e42-4451-80ee-6da60aeceacd
        type: string
      timestamp:
        example: "2026-01-19T15:37:27.514667373Z"
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        example: auth_failed
        type: string
      details: {}
      message:
        example: Authentication failed
        type: string
      request_id:
        example: 7fbd6854-8e42-4451-80ee-6da60aeceacd
        type: string
      timestamp:
        example: "2026-01-19T15:37:27.514667373Z"
        type: string
    type: object
...
  response.ShiftDoc:
    properties:
      created_at:
        example: "2026-01-19T15:37:27.514667373Z"
        type: string
      end_time:
        example: "2026-03-22T18:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      note:
        example: Front desk day shift
        type: string
      start_time:
        example: "2026-03-22T09:00:00Z"
        type: string
      updated_at:
        example: "2026-01-19T15:37:27.514667373Z"
        type: string
      user:
        $ref: '#/definitions/response.ShiftUserDoc'
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      zone_tag_id:
        example: 1
        type: integer
    type: object
  response.ShiftListResponse:
    properties:
      request_id:
        example: 7fbd6854-8e42-4451-80ee-6da60aeceacd
        type: string
      shifts:
        items:
          $ref: '#/definitions/response.ShiftDoc'
        type: array
      timestamp:
        example: "2026-01-19T15:37:27.514667373Z"
        type: string
    type: object
  response.ShiftResponse:
    properties:
      request_id:
        example: 7fbd6854-8e42-4451-80ee-6da60aeceacd
        type: string
      shift:
        $ref: '#/definitions/response.ShiftDoc'
      timestamp:
        example: "2026-01-19T15:37:27.514667373Z"
        type: string
    type: object
  response.ShiftUserDoc:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      email:
        example: ada@example.com
        type: string
      full_name:
        example: Ada Lovelace
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      phone:
        example: "+79991234567"
        type: string
      role:
        example: admin
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: Backend API for PlayGround
  title: PlayGround API
  version: "1.0"
paths:
...
  /api/v1/shift:
    get:
      description: Returns all shifts with creator user data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ShiftListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List shifts
      tags:
      - shifts
    post:
      consumes:
      - application/json
      description: Creates a shift from JSON payload and attaches the current authenticated
        admin user
      parameters:
      - description: Shift payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.CreateShiftRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ShiftResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create shift
      tags:
      - shifts
  /api/v1/shift/{id}:
    delete:
      description: Soft-deletes shift by id
      parameters:
      - description: Shift ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DeleteShiftResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete shift
      tags:
      - shifts
    get:
      description: Returns shift by id with creator user data
      parameters:
      - description: Shift ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ShiftResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get shift by id
      tags:
      - shifts
    patch:
      consumes:
      - application/json
      description: Partially updates shift fields by id
      parameters:
      - description: Shift ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Shift patch payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.PatchShiftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ShiftResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Patch shift
      tags:
      - shifts
  /api/v1/shift/zone-tag/{zone_tag_id}:
    get:
      description: Returns all upcoming and active shifts for a given zone_tag_id
      parameters:
      - description: Zone Tag ID
        format: int64
        in: path
        name: zone_tag_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ShiftListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get shifts by zone tag id
      tags:
      - shifts
...
  /docs:
    get:
      description: Returns the Scalar documentation page for admin users
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: file
        "301":
          description: Redirect to frontend
          schema:
            type: string
      summary: Get Tool Docs
      tags:
      - Tool
...
  /openapi.json:
    get:
      description: Returns the generated OpenAPI specification for admin users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "301":
          description: Redirect to frontend
          schema:
            type: string
      summary: Get OpenAPI Spec
      tags:
      - Tool
...
swagger: "2.0"